Move e2e client into internal/
